+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 ## [Unreleased]
+
+### Improvements
+
+- In Latex, parse `\operatorname{foo}` as the MathJSON symbol `"foo"`.